# api/openapi.yaml
openapi: 3.0.3
info:
  title: Mie & Beurre Bakery API
  description: Backend API for the Mie & Beurre bakery marketplace application.
  version: 1.0.0
  contact:
    name: Mie & Beurre Team
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.mieetbeurre.com
    description: Production
...
security:
  - BearerAuth: []
paths:
...
  /api/bakeries/{id}/reviews:
    get:
      tags: [Reviews]
      summary: List bakery reviews
      description: Returns a paginated list of reviews for a bakery (public).
      security: []
      parameters:
        - $ref: '#/components/parameters/BakeryId'
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Paginated list of reviews
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewListResponse'

    post:
      tags: [Reviews]
      summary: Create a review
      description: Submit a review for a bakery. Requires a delivered order at this bakery.
      parameters:
        - $ref: '#/components/parameters/BakeryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '201':
          description: Review created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '400':
          description: Invalid rating or text too long
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not a verified purchaser
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Review already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/reviews/{id}/report:
    post:
      tags: [Reviews]
      summary: Report a review
      description: Flag a review for moderation.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportReviewRequest'
      responses:
        '204':
          description: Review reported
        '404':
          description: Review not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/seller/reviews/{id}/hide:
    put:
      tags: [Reviews]
      summary: Hide a review
      description: Bakery owner hides a review from public display (seller only).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Review hidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '403':
          description: Forbidden - not the bakery owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Review not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─── B2B Comptoir ───────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    BakeryId:
      name: id
      in: path
      required: true
      description: Bakery ID
      schema:
        type: string
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: INVALID_REQUEST
        message:
          type: string
          example: "invalid request body"
...
    CreateReviewRequest:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        text:
          type: string
          maxLength: 1000
    ReviewResponse:
      type: object
      properties:
        id:
          type: string
        rating:
          type: integer
        text:
          type: string
          nullable: true
        authorName:
          type: string
        createdAt:
          type: string
          format: date-time
    ReportReviewRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
    ReviewListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReviewResponse'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer

    # ─── B2B ────────────────────────────────────────────────────────────────
...